+++ tests/test_tool_call.py
@@ -153,7 +153,7 @@ def test_simple_toolset():
     toolset = SimpleToolset([PlusTool()])
     toolset += CompareTool()
     toolset += RaiseTool()
-    toolset += ErrorTool()
+    toolset.add(ErrorTool())
     assert toolset.tools[0].name == "plus"
     assert toolset.tools[1].name == "compare"
     assert toolset.tools[2].name == "raise"
@@ -272,3 +272,22 @@ def test_callable_tool_2():
     )
     assert asyncio.run(tool.call({"a": 1})) == ToolOk(output="Test tool called with 1 and 0")
     assert isinstance(asyncio.run(tool.call({"b": 2})), ToolValidateError)
+
+
+def test_simple_toolset_sub():
+    class TestParams(BaseModel):
+        pass
+
+    class TestTool(CallableTool2[TestParams]):
+        name: str = "test"
+        description: str = "This is a test tool"
+        params: type[TestParams] = TestParams
+
+        @override
+        async def __call__(self, params: TestParams) -> ToolReturnValue:
+            return ToolOk(output="Test tool called")
+
+    toolset = SimpleToolset([TestTool()])
+    assert len(toolset.tools) == 1
+    toolset.remove(TestTool.name)
+    assert len(toolset.tools) == 0